+++ src/app/api/exhibitions/[exhibitionId]/reviews/route.ts
@@ -1,7 +1,7 @@
 import { NextRequest, NextResponse } from 'next/server';
 import { createClient } from '@/lib/supabase/server';
 import { reviewCreateSchema } from '@/lib/schemas/review';
-import { fetchExhibitionReviews } from '@/lib/exhibition/queries';
+import { fetchExhibitionReviews } from '@/lib/review/server';
 
 type RouteContext = { params: Promise<{ exhibitionId: string }> };
