+++ src/components/home/ExhibitionListContent.tsx
@@ -1,4 +1,4 @@
-import { fetchExhibitions } from '@/lib/exhibition/queries';
+import { fetchExhibitions } from '@/lib/exhibition/server';
 import ExhibitionList from './ExhibitionList';
 import ListPagination from './ListPagination';
 import { ExhibitionSort } from '@/types/exhibition';
