+++ src/lib/auth/getAuthContext.ts
@@ -1,3 +1,5 @@
+import 'server-only';
+
 import { cache } from 'react';
 import type { User } from '@supabase/supabase-js';
 import { createClient } from '@/lib/supabase/server';
