+++ src/lib/exhibition/constants.ts
@@ -1,2 +1,1 @@
 export const EXHIBITIONS_PER_PAGE = 12;
-export const REVIEWS_PER_PAGE = 10;
