+++ src/lib/review/constants.ts
@@ -1,0 +1,1 @@
+export const REVIEWS_PER_PAGE = 10;
