+++ src/lib/review/service.ts
@@ -1,64 +1,5 @@
 import { ExhibitionReviewItem } from '@/types/exhibition';
 
-export const postNewExhibition = async (formDate: FormData) => {
-  const res = await fetch('/api/exhibitions', {
-    method: 'POST',
-    body: formDate,
-  });
-
-  if (!res.ok) {
-    const error = await res.json();
-    throw new Error(error.message);
-  }
-  const { createdId } = await res.json();
-
-  return createdId;
-};
-export const getExhibitionDetails = async (id: string) => {
-  const res = await fetch(`/api/exhibitions/${id}`);
-
-  if (!res.ok) {
-    const error = await res.json();
-    throw new Error(error.message);
-  }
-  const { data } = await res.json();
-  const title = data?.title || '';
-  const host = data?.host || '';
-  return { title, host };
-};
-export const toggleExhibitionLike = async (
-  exhibitionId: string,
-  nextLiked: boolean
-) => {
-  const res = await fetch(`/api/exhibitions/${exhibitionId}/likes`, {
-    method: nextLiked ? 'POST' : 'DELETE',
-  });
-
-  if (!res.ok) {
-    const error = await res.json().catch(() => ({}));
-    throw new Error(error.message ?? 'like error');
-  }
-
-  return res.json();
-};
-
-export const endExhibition = async (exhibitionId: string) => {
-  const res = await fetch(`/api/exhibitions/${exhibitionId}`, {
-    method: 'PATCH',
-    body: JSON.stringify({
-      end_date: new Date(),
-    }),
-  });
-
-  if (!res.ok) {
-    const error = await res.json();
-    throw new Error(error.message);
-  }
-  const { updatedId } = await res.json();
-
-  return updatedId;
-};
-
 export interface ReviewsPage {
   reviews: ExhibitionReviewItem[];
   totalCount: number;
